requirements.txt
Flask
kafka-python
psycopg2
json
logging
